Added the Database Link Code
The correct database driver is selected on the mDatabase field
initialization with
the QSqlDatabase::addDatabase("QSQLITE") function call. The following
steps are just a matter of
configuring the database name (which is incidentally the file path in
SQLITE ) and opening the
connection with the mDatabase->open() function. In the DatabaseManager
destructor, the connection is
closed and the mDatabase pointer is properly deleted.